requirements.txt
opencv-python>=4.8.0
numpy>=1.24.3
tensorflow>=2.13.0
mediapipe>=0.10.0
pandas>=2.0.3
python-dotenv>=1.0.0
playsound>=1.3.0 